requirements.txt
pytesseract
pillow
mirascope
groq
google-generativeai
streamlit
pyMuPDF
scikit-learn
transformers
torch
nltk
pandas
tenacity
